fix: make interactive TUI run without panicking (darwin)
Two crashes hit on launch/scan; both are faithful-port fixes:

1. checkRecalculation (ScreenManager.c:152-153): the port dropped the
   'always update header' Header_updateData call that C runs before Header_draw.
   Without it the header meters were drawn at their Meter_new init state
   (curItems == maxItems, e.g. CPU's 10) which over-ran the shorter class
   attributes palette (8) -> index-out-of-bounds in BarMeterMode_draw. Restored
   the call before the redraw block.

2. Process_updateCPUFieldWidths (Process.c:1122): C computes
   ceil(log10(pct+0.1))+2 in floating point and truncates to a uint8_t width.
   The port cast to i32 mid-expression, so a non-finite pct (the first darwin
   sample's zero-delta CPU%) saturated to i32::MAX and then '+2' overflowed in
   debug. Keep the expression in f32 and cast to u8 last, matching the C type.

Verified via pty: renders the full header (18 CPU meters, Mem/Swp, Tasks) and
process list, and quits cleanly on 'q' (exit 0), no panics.

+++ src/ported/process.rs
@@ -2357,8 +2357,13 @@ pub fn Process_updateCPUFieldWidths(percentage: f32) {
         return;
     }
 
-    // Two extra characters: one for "." and another for precision.
-    let width = ((percentage + 0.1).log10().ceil() as i32 + 2) as usize;
+    // Two extra characters: one for "." and another for precision. C computes
+    // `ceil(log10(percentage + 0.1)) + 2` in floating point and truncates to
+    // the `uint8_t width`; keep the whole expression in `f32` and cast to `u8`
+    // last (matching the C type) so a non-finite `percentage` — e.g. the first
+    // darwin sample's zero-delta CPU% — saturates to `u8::MAX` instead of
+    // overflowing an intermediate `i32 + 2`.
+    let width = ((percentage + 0.1).log10().ceil() + 2.0) as u8 as usize;
     Row_updateFieldWidth(ProcessField::PERCENT_CPU as RowField, width);
     Row_updateFieldWidth(ProcessField::PERCENT_NORM_CPU as RowField, width);
 }
